#38: restructure project like #43
