+++ lib/backup/index.js
@@ -8,8 +8,8 @@ const {
   fetchSingleDDReport,
   fetchDataAvailabilityReport,
   sleep
-} = require('../../data-access/cert-api-client');
-const { multiUsiUoiMappings } = require('../../data-access/usi-mapping');
+} = require('../misc/data-access/cert-api-client');
+const { multiUsiUoiMappings } = require('../misc/data-access/usi-mapping');
 const { ENDORSEMENTS_PATH } = process.env;
 
 const BACKUP_DIRECTORY = 'reso-server-backup',
