+++ lib/batch-test-runner/index.js
