+++ lib/restore-utils/index.js
@@ -8,7 +8,7 @@ const {
   getOrgsMap,
   processDataDictionaryResults,
   restoreBackedUpReport
-} = require('../../data-access/cert-api-client');
+} = require('../misc/data-access/cert-api-client');
 const { isValidUrl, fetchSystemData } = require('../../common');
 
 const { processLookupResourceMetadataFiles } = require('reso-certification-etl');
