+++ lib/transform/index.js
@@ -8,7 +8,7 @@ const {
   fetchDataAvailabilityReport,
   sleep,
   postTransformedReport
-} = require('../../data-access/cert-api-client');
+} = require('../misc/data-access/cert-api-client');
 const { backup: createBackup } = require('../backup');
 
 const composeTransformations = async (transformations = [], initialData = {}) => {
