+++ lib/transform/transformations/rescore.js
